update xóa thành viên

See merge request tuanphanminhanh/nmcnpm.20221-nhom4!68

+++ ProjectCNPM/src/main/java/QuanLyNhanKhau/controllers/COVIDController.java
@@ -10,10 +10,7 @@ import javafx.fxml.FXML;
 import javafx.fxml.Initializable;
 import javafx.scene.Parent;
 import javafx.scene.Scene;
-import javafx.scene.control.Button;
-import javafx.scene.control.TableColumn;
-import javafx.scene.control.TableView;
-import javafx.scene.control.TextField;
+import javafx.scene.control.*;
 import javafx.scene.control.cell.PropertyValueFactory;
 import javafx.scene.layout.BorderPane;
 import javafx.stage.Stage;
@@ -21,6 +18,7 @@ import javafx.stage.Stage;
 import java.io.IOException;
 import java.net.URL;
 import java.sql.SQLException;
+import java.util.Optional;
 import java.util.ResourceBundle;
 
 public class COVIDController implements Initializable {
@@ -81,8 +79,15 @@ public class COVIDController implements Initializable {
         } else if (event.getSource() == xoaNguoiMac) {
             CovidTable tmp = table.getSelectionModel().getSelectedItem();
             if (tmp != null) {
-                covidDB.deleteMacCovid(tmp.getId());
-                table.getItems().remove(tmp);
+                Alert alert = new Alert(Alert.AlertType.CONFIRMATION);
+                alert.setTitle("Xác nhận");
+                alert.setHeaderText(null);
+                alert.setContentText("Bạn có muốn xóa không?");
+                Optional<ButtonType> result = alert.showAndWait();
+                if (result.get() == ButtonType.OK) {
+                    covidDB.deleteMacCovid(tmp.getId());
+                    table.getItems().remove(tmp);
+                }
             }
         } else if (event.getSource() == capNhatTT) {
             Parent root = Windows.getRoot("covid/capnhatnguoimac.fxml");
@@ -96,7 +101,7 @@ public class COVIDController implements Initializable {
             ObservableList<CovidTable> list_search = FXCollections.observableArrayList();
             String inputHoTen = hoTenInput.getText();
             for (CovidTable covidTable : listCovid) {
-                if (covidTable.getHoTen().contains(inputHoTen)) {
+                if (covidTable.getHoTen().toLowerCase().contains(inputHoTen.toLowerCase())) {
                     list_search.add(covidTable);
                 }
             }
